Revamp of my initial blog project. I added responsive design.a simple 'design' that i will maybe update in the futur.
You can now send a news which contain an optionnal Url for mP3 (use in my IOS app) An url with a video of the song, upload at the same time a picture which illustrate the new, a title and a little description.

I did too refactoring of my previous code. When you post a news now the code generate a Json which will be use in my IOS application that you can find on my repository.